last bit of cleanup around docs and some last tweaks
