+++ quixstreams/sources/community/file/file.py
@@ -21,6 +21,7 @@ class FileSource(Source):
 
     Expects folder and file structures as generated by the related FileSink connector:
 
+    ```
     my_topics/
     ├── topic_a/
     │   ├── 0/
@@ -31,6 +32,7 @@ class FileSource(Source):
     │       └── 0016.ext
     └── topic_b/
         └── etc...
+    ```
 
     Intended to be used with a single topic (ex: topic_a), but will recursively read
     from whatever entrypoint is passed to it.
@@ -39,8 +41,9 @@ class FileSource(Source):
 
     See the `.formats` and `.compressions` modules to see what is supported.
 
-    Example:
+    Example Usage:
 
+    ```python
     from quixstreams import Application
     from quixstreams.sources.community.file import FileSource
 
@@ -54,6 +57,7 @@ class FileSource(Source):
 
     if __name__ == "__main__":
         app.run()
+    ```
     """
 
     def __init__(
