+++ quixstreams/sources/community/gcp_pubsub/__init__.py
@@ -1,2 +1,3 @@
 # ruff: noqa: F403
+from .components import *
 from .gcp_pubsub import *
