+++ quixstreams/sources/community/gcp_pubsub/components/__init__.py
@@ -1,0 +1,2 @@
+# ruff: noqa: F403
+from .config import *
